% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AoptimalGauLMSub.R
\name{AoptimalGauLMSub}
\alias{AoptimalGauLMSub}
\title{A-optimality criteria based subsampling under Gaussian Linear Models}
\usage{
AoptimalGauLMSub(r1,r2,Y,X,N)
}
\arguments{
\item{r1}{sample size for initial random sampling}

\item{r2}{sample size for optimal sampling}

\item{Y}{response data or Y}

\item{X}{covariate data or X matrix that has all the covariates (first column is for the intercept)}

\item{N}{size of the big data}
}
\value{
The output of \code{AoptimalGauLMSub} gives a list of

\code{Beta_Estimates} estimated model parameters in a data.frame after subsampling

\code{Variance_Epsilon_Estimates} matrix of estimated variance for epsilon in a data.frame after subsampling

\code{Sample_A-Optimality} list of indexes for the initial and optimal samples obtained based on A-Optimality criteria

\code{Subsampling_Probability} matrix of calculated subsampling probabilities for A-optimality criteria
}
\description{
Using this function sample from big data under Gaussian linear regression models
to describe the data. Subsampling probabilities are obtained based on the A-optimality criteria.
}
\details{
Two stage subsampling algorithm for big data under Gaussian Linear Model.

First stage is to obtain a random sample of size \eqn{r_1} and estimate the model parameters.
Using the estimated parameters subsampling probabilities are evaluated for A-optimality criteria.

Through the estimated subsampling probabilities an optimal sample of size \eqn{r_2 \ge r_1} is obtained.
Finally, the two samples are combined and the model parameters are estimated.

\strong{NOTE} : If input parameters are not in given domain conditions
necessary error messages will be provided to go further.

If \eqn{r_2 \ge r_1} is not satisfied then an error message will be produced.

If the big data \eqn{X,Y} has any missing values then an error message will be produced.

The big data size \eqn{N} is compared with the sizes of \eqn{X,Y} and if they are not aligned an error
message will be produced.
}
\examples{
Dist<-"Normal"; Dist_Par<-list(Mean=0,Variance=1,Error_Variance=0.5)
No_Of_Var<-2; Beta<-c(-1,2,1); N<-10000; Family<-"linear"
Full_Data<-GenGLMdata(Dist,Dist_Par,No_Of_Var,Beta,N,Family)

r1<-300; r2<-rep(100*c(6,12),50); Original_Data<-Full_Data$Complete_Data;

AoptimalGauLMSub(r1 = r1, r2 = r2,
                 Y = as.matrix(Original_Data[,colnames(Original_Data) \%in\% c("Y")]),
                 X = as.matrix(Original_Data[,-1]),
                 N = nrow(Original_Data))->Results

plot_Beta(Results)

}
\references{
\insertRef{lee2021fast}{NeEDS4BigData}
}
