% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_data.R
\docType{data}
\name{Bike_sharing}
\alias{Bike_sharing}
\title{Bike sharing data}
\format{
A data frame with 4 columns and 17,379 rows.
\describe{
\item{\code{Rented_Bikes}}{Number of bikes rented hourly}
\item{\code{Temperature}}{Hourly temperature}
\item{\code{Humidity}}{Hourly humidity}
\item{\code{Windspeed}}{Hourly windspeed}
}
}
\source{
Extracted from

Fanaee-T H (2013) Bike Sharing. UCI Machine Learning Repository.

Available at: \doi{10.24432/C5W894}
}
\usage{
Bike_sharing
}
\description{
Fanaee-T (2013) collected data to understand the bike sharing demands under
the rental and return process. The data in total contains 17,379 observations
where we consider the covariates temperature, humidity and windspeed
to model the response, the number of bikes rented hourly.
}
\examples{
nrow(Bike_sharing)

}
\keyword{datasets}
