% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_data.R
\docType{data}
\name{One_million_songs}
\alias{One_million_songs}
\title{One million songs data}
\format{
A data frame with 4 columns and 309,685 rows.
\describe{
\item{\code{Counts}}{Number of playback counts for songs}
\item{\code{Duration}}{Duration of the song}
\item{\code{Loudness}}{Loudness of the song}
\item{\code{Tempo}}{Tempo of the song}
\item{\code{Artist_Hotness}}{A value between 0 and 1}
\item{\code{Song_Hotness}}{A value between 0 and 1}
}
}
\usage{
One_million_songs
}
\description{
This data set contains 1,019,318 unique users' music play counts in the Echo Nest,
which is available at "http://millionsongdataset.com/tasteprofile/".
As a basic step, it is interesting to predict the play counts using the song
information collected in the Million Song Dataset (Bertin-Mahieux et al. (2011)).
After cleaning up and feature engineering the data in total contains 205,032
observations where we consider the covariates duration, loudness, tempo, artist
hotness and song hotness to model the response,
the number of song counts.
}
\examples{
nrow(One_million_songs)

}
\references{
\insertRef{mcfee2012million}{NeEDS4BigData}

\insertRef{ai2021optimal}{NeEDS4BigData}
}
\keyword{datasets}
