% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_LmAMSE}
\alias{plot_LmAMSE}
\title{Plotting loss function or mAMSE outputs for the subsamples under model misspecification}
\usage{
plot_LmAMSE(object)
}
\arguments{
\item{object}{Any object after subsampling from our subsampling function under potential model misspecification}
}
\value{
The output is a faceted ggplot result
}
\description{
After using the subsampling methods under potential model misspecification we obtain
their respective loss or mAMSE values. They are summarised as plots here.
}
\details{
For A- and L-optimality criterion and LmAMSE subsampling under Generalised Linear Models
with potential model misspecification the facets are for variance and bias^2 of mAMSE values.
}
