% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Utility}
\alias{plot_Utility}
\title{Plotting A- and D-optimality for model parameter (Beta) outputs of the subsamples}
\usage{
plot_Utility(object)
}
\arguments{
\item{object}{Any object after subsampling from our subsampling functions}
}
\value{
The output is a faceted ggplot result
}
\description{
After using the subsampling methods we mostly obtain the estimated model parameter
estimates. Of these estimates we can obtain their respective A- and D-optimality
values and they are summarised as plots here. It should be noted that
A- and D-optimality represents the model parameter's tr(Variance) and log(det(Information)),
respectively.
}
\details{
For A- and L-optimality criterion subsampling under Generalised Linear Models
the facets are for model parameter variance and information.

For A- and L-optimality criterion subsampling under Generalised Linear Models
where multiple models can describe the data the facets are for model parameter variance and information.
}
