% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_data.R
\docType{data}
\name{Skin_segmentation}
\alias{Skin_segmentation}
\title{Skin segmentation data}
\format{
A data frame with 4 columns and 245,057 rows.
\describe{
\item{\code{Skin_presence}}{Skin presence in the randomly selected pixels}
\item{\code{Red}}{Red values in the randomly selected pixels}
\item{\code{Green}}{Green values in the randomly selected pixels}
\item{\code{Blue}}{Blue values in the randomly selected pixels}
}
}
\source{
Extracted from

Rajen B, Abhinav D (2012) Skin segmentation. UCI Machine Learning Repository.

Available at: \doi{10.24432/C5T30C}
}
\usage{
Skin_segmentation
}
\description{
Rajen and Abhinav (2012) addressed the challenge of detecting skin-like
regions in images as a component of the intricate process of facial recognition.
To achieve this goal, they curated the “Skin segmentation” data set, comprising
RGB (R-red, G-green, B-blue) values of randomly selected pixels from
N = 245,057 facial images, including 50,859 skin samples and 194,198 nonskin
samples, spanning diverse age groups, racial backgrounds, and genders.
}
\examples{
nrow(Skin_segmentation)

}
\keyword{datasets}
