% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMissLinSub.R
\name{GenModelMissGLMdata}
\alias{GenModelMissGLMdata}
\title{Generate data for Generalised Linear Models under model misspecification scenario}
\usage{
GenModelMissGLMdata(No_Of_Var,Beta,Var_Epsilon,N,MisspecificationType,family)
}
\arguments{
\item{No_Of_Var}{number of variables}

\item{Beta}{a vector for the model parameters, including the intercept}

\item{Var_Epsilon}{variance value for the residuals}

\item{N}{the big data size}

\item{MisspecificationType}{a character vector referring to different types of misspecification}

\item{family}{a character vector for "linear", "logistic" and "poisson" regression from Generalised Linear Models}
}
\value{
The output of \code{GenModelMissGLMdata} gives a list of

\code{N} the big data size
\code{Beta} a list of outputs(real and estimated) for the beta values
\code{Variance_Epsilon} a list of outputs(real and estimated) for the variance epsilon
\code{Xbeta} a list of outputs(real and estimated) for the linear predictor
\code{f} a list of outputs(real and estimated) misspecification
\code{Real_Full_Data} a matrix for Y,X and f(x)
\code{Full_Data} a matrix for Y and X
}
\description{
Function to simulate big data under Generalised Linear Models for the model misspecification scenario through
any misspecification type.
}
\details{
Big data for the Generalised Linear Models are generated by the "linear", "logistic" and "poisson"
regression types under model misspecification.

We have limited the covariate data generation through uniform distribution of limits \eqn{(-1,1)}.

Different type of misspecifications are "Type 1", "Type 2 Squared", "Type 2 Interaction",
"Type 3 Squared" or "Type 3 Interaction".
}
\examples{
No_Of_Var<-2; Beta<-c(-1,2,2,1); Var_Epsilon<-0.5; N<-10000;
MisspecificationType <- "Type 2 Squared"; family <- "linear"

Results<-GenModelMissGLMdata(No_Of_Var,Beta,Var_Epsilon,N,MisspecificationType,family)

No_Of_Var<-2; Beta<-c(-1,2,2,1); N<-10000;
MisspecificationType <- "Type 2 Squared"; family <- "logistic"

Results<-GenModelMissGLMdata(No_Of_Var,Beta,Var_Epsilon=NULL,N,MisspecificationType,family)

No_Of_Var<-2; Beta<-c(-1,2,2,1); N<-10000;
MisspecificationType <- "Type 2 Squared"; family <- "poisson"

Results<-GenModelMissGLMdata(No_Of_Var,Beta,Var_Epsilon=NULL,N,MisspecificationType,family)

}
