% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMissPoiSub.R
\name{modelMissPoiSub}
\alias{modelMissPoiSub}
\title{Subsampling under Poisson regression for a potentially misspecified model}
\usage{
modelMissPoiSub(r1,r2,Y,X,N,Alpha,Beta_Estimate_Full,F_Estimate_Full)
}
\arguments{
\item{r1}{subsample size for initial random sampling}

\item{r2}{subsample size for optimal sampling}

\item{Y}{response data or Y}

\item{X}{covariate data or X matrix that has all the covariates (first column is for the intercept)}

\item{N}{size of the big data}

\item{Alpha}{scaling factor when using Log Odds or Power functions to magnify the probabilities}

\item{Beta_Estimate_Full}{estimate of Beta after fitting the Poisson model}

\item{F_Estimate_Full}{estimate of f that is the difference of linear predictor on GAM and Poisson model}
}
\value{
The output of \code{modelMissPoiSub} gives a list of

\code{Beta_Estimates} estimated model parameters after subsampling

\code{AMSE_Estimates} matrix of estimated AMSE values after subsampling

\code{Sample_A-Optimality} list of indexes for the initial and optimal subsamples obtained based on A-Optimality criterion

\code{Sample_L-Optimality} list of indexes for the initial and optimal subsamples obtained based on L-Optimality criterion

\code{Sample_RLmAMSE} list of indexes for the optimal subsamples obtained based on RLmAMSE

\code{Sample_RLmAMSE_Log_Odds} list of indexes for the optimal subsamples obtained based on RLmAMSE with Log Odds function

\code{Sample_RLmAMSE_Power} list of indexes for the optimal subsamples obtained based on RLmAMSE with Power function

\code{Subsampling_Probability} matrix of calculated subsampling probabilities
}
\description{
Using this function subsample from big data under Poisson regression for a potentially misspecified model.
Subsampling probabilities are obtained based on the A- and L- optimality criteria
with the RLmAMSE (Reduction of Loss by minimizing the Average Mean Squared Error).
}
\details{
Two stage subsampling algorithm for big data under Poisson regression for potential model misspecification.

First stage is to obtain a random sample of size \eqn{r_1} and estimate the model parameters.
Using the estimated parameters subsampling probabilities are evaluated for A-, L-optimality criterion,
RLmAMSE and enhanced RLmAMSE (log-odds and power) subsampling methods.

Through the estimated subsampling probabilities an subsample of size \eqn{r_2 \ge r_1} is obtained.
Finally, the two samples are combined and the model parameters are estimated for A- and L-optimality,
while for RLmAMSE and enhanced RLmAMSE (log-odds and power) only the optimal subsample is used.

\strong{NOTE} :  If input parameters are not in given domain conditions
necessary error messages will be provided to go further.

If \eqn{r_2 \ge r_1} is not satisfied then an error message will be produced.

If the big data \eqn{X,Y} has any missing values then an error message will be produced.

The big data size \eqn{N} is compared with the sizes of \eqn{X,Y},F_estimate_Full and
if they are not aligned an error message will be produced.

If \eqn{\alpha > 1} for the scaling vector is not satisfied an error message will be produced.
}
\examples{
No_Of_Var<-2; Beta<-c(-1,2,2,1); N<-10000;
MisspecificationType <- "Type 2 Squared"; family <- "poisson"

Full_Data<-GenModelMissGLMdata(No_Of_Var,Beta,Var_Epsilon=NULL,N,MisspecificationType,family)

r1<-300; r2<-100*c(6,9); Original_Data<-Full_Data$Full_Data;

# cl <- parallel::makeCluster(4)
# doParallel::registerDoParallel(cl)
\dontrun{
Results<-modelMissPoiSub(r1 = r1, r2 = r2,
                         Y = as.matrix(Original_Data[,1]),
                         X = as.matrix(Original_Data[,-1]),
                         N = Full_Data$N,
                         Alpha = 10,
                         Beta_Estimate_Full = Full_Data$Beta$Estimate,
                         F_Estimate_Full = Full_Data$f$Real_GAM)

# parallel::stopCluster(cl)

plot_Beta(Results)
plot_AMSE(Results)
}

}
\references{
\insertRef{adewale2009robust}{NeEDS4BigData}
\insertRef{adewale2010robust}{NeEDS4BigData}
}
