% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelRobustLogSub.R
\name{modelRobustLogSub}
\alias{modelRobustLogSub}
\title{Model robust optimal subsampling for A- and L- optimality criterion under logistic regression}
\usage{
modelRobustLogSub(r1,r2,Y,X,N,Alpha,All_Combinations,All_Covariates)
}
\arguments{
\item{r1}{subsample size for initial random sampling}

\item{r2}{subsample size for optimal sampling}

\item{Y}{response data or Y}

\item{X}{covariate data or X matrix that has all the covariates (first column is for the intercept)}

\item{N}{size of the big data}

\item{Alpha}{vector of alpha values that are used to obtain the model robust subsampling probabilities}

\item{All_Combinations}{list of possible models that can describe the data}

\item{All_Covariates}{all the covariates in the models}
}
\value{
The output of \code{modelRobustLinSub} gives a list of

\code{Beta_Data} estimated model parameters for each model in a list after subsampling

\code{Utility_Data} estimated Variance and Information of the model parameters after subsampling

\code{Sample_L-optimality} list of indexes for the initial and optimal subsamples obtained based on L-optimality criterion

\code{Sample_L-optimality_MR} list of indexes for the initial and model robust optimal subsamples obtained based on L-optimality criterion

\code{Sample_A-optimality} list of indexes for the initial and optimal subsamples obtained based on A-optimality criterion

\code{Sample_A-optimality_MR} list of indexes for the initial and model robust optimal subsamples obtained based on A-optimality criterion

\code{Subsampling_Probability} matrix of calculated subsampling probabilities for A- and L- optimality criterion
}
\description{
Using this function subsample from big data under logistic regression when there are more than
one model to describe the data. Subsampling probabilities are obtained based on the A- and L-
optimality criterions.
}
\details{
Two stage subsampling algorithm for big data under logistic regression for multiple models that can
describe the big data.

First stage is to obtain a random sample of size \eqn{r_1} and estimate the model parameters for all models.
Using the estimated parameters subsampling probabilities are evaluated for A-, L-optimality criterion and
model averaging A-, L-optimality subsampling methods.

Through the estimated subsampling probabilities an subsample of size \eqn{r_2 \ge r_1} is obtained.
Finally, the two samples are combined and the model parameters are estimated for all the models.

\strong{NOTE} :  If input parameters are not in given domain conditions
necessary error messages will be provided to go further.

If \eqn{r_2 \ge r_1} is not satisfied then an error message will be produced.

If the big data \eqn{X,Y} has any missing values then an error message will be produced.

The big data size \eqn{N} is compared with the sizes of \eqn{X,Y} and
if they are not aligned an error message will be produced.

If \eqn{0 < \alpha < 1} for the a priori probabilities are not satisfied an error message will be produced.
}
\examples{
Dist<-"Normal"; Dist_Par<-list(Mean=0,Variance=1)
No_Of_Var<-2; Beta<-c(-1,2,1,2); N<-10000
All_Models<-list(Real_Model=c("X0","X1","X2","X1^2"),
                 Assumed_Model_1=c("X0","X1","X2"),
                 Assumed_Model_2=c("X0","X1","X2","X2^2"),
                 Assumed_Model_3=c("X0","X1","X2","X1^2","X2^2"))
family = "logistic"

Full_Data<-GenModelRobustGLMdata(Dist,Dist_Par,No_Of_Var,Beta,N,All_Models,family)

r1<-300; r2<-rep(100*c(6,9,12),25); Original_Data<-Full_Data$Complete_Data;

modelRobustLogSub(r1 = r1, r2 = r2,
                  Y = as.matrix(Original_Data[,colnames(Original_Data) \%in\% c("Y")]),
                  X = as.matrix(Original_Data[,-1]),N = nrow(Original_Data),
                  Alpha = rep(1/length(All_Models),length(All_Models)),
                  All_Combinations = All_Models,
                  All_Covariates = colnames(Original_Data)[-1])->Results

Beta_Plots<-plot_Beta(Results)

}
\references{
\insertRef{mahendran2023model}{NeEDS4BigData}
}
