% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Beta}
\alias{plot_Beta}
\title{Plotting model parameter outputs after subsampling}
\usage{
plot_Beta(object)
}
\arguments{
\item{object}{Any object after subsampling from our subsampling functions}
}
\value{
The output is a faceted ggplot result
}
\description{
After using the subsampling methods we mostly obtain the estimated model parameter
estimates. Here, they are summarised as histogram plots.
}
\details{
For local case control sampling the facets are for subsample sizes and beta values.

For leverage sampling the facets are for subsample sizes and beta values.

For A- and L-optimality criteria subsampling under Generalised Linear Models
the facets are for subsample sizes and beta values.

For A-optimality criteria subsampling under Gaussian Linear Models
the facets are for subsample sizes and beta values.

For A-optimality criteria subsampling under Generalised Linear Models
with response variable not inclusive the facets are for subsample sizes and beta values.

For A- and L-optimality criteria subsampling under Generalised Linear Models
where multiple models can describe the data the facets are for subsample sizes and beta values.

For A- and L-optimality criteria and LmAMSE subsampling under Generalised Linear Models
with potential model misspecification the facets are for subsample sizes and beta values.
}
